Apply new PackageContext in package and better preview
